# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Quotation App
  version: 2.0.0
  description: Multi-tenant SaaS quotation & CRM platform for MSMEs
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /leads/{id}/score:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      tags: [Leads]
      summary: Re-score lead (AI + rules)
      operationId: scoreLead
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lead" }
...
  /invoices:
    get:
      tags: [Invoices]
      operationId: listInvoices
      parameters:
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invoice" }
    post:
      tags: [Invoices]
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [Invoices]
      operationId: getInvoice
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceDetail" }
    patch:
      tags: [Invoices]
      operationId: updateInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/{id}/status:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      tags: [Invoices]
      operationId: setInvoiceStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
  /invoices/from-sales-order/{salesOrderId}:
    parameters:
      - in: path
        name: salesOrderId
        required: true
        schema: { type: integer }
    post:
      tags: [Invoices]
      operationId: promoteSalesOrderToInvoice
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }

  # ── Payments ──
...
  /integrations/{provider}:
    parameters:
      - in: path
        name: provider
        required: true
        schema: { type: string }
    put:
      tags: [Integrations]
      operationId: upsertIntegration
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IntegrationInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Integration" }
    delete:
      tags: [Integrations]
      operationId: deleteIntegration
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Auth ──
...
    Lead:
      type: object
      required: [id, name, source, status, priority, score, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        source: { type: string }
        externalId: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        score: { type: integer }
        budget: { type: ["number", "null"] }
        product: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        nextAction: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
        convertedClientId: { type: ["integer", "null"] }
        lastContactedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    Invoice:
      type: object
      required: [id, invoiceNumber, status, total, createdAt]
      properties:
        id: { type: integer }
        invoiceNumber: { type: string }
        clientId: { type: ["integer", "null"] }
        clientName: { type: ["string", "null"] }
        salesOrderId: { type: ["integer", "null"] }
        status: { type: string }
        issueDate: { type: string }
        dueDate: { type: ["string", "null"] }
        sellerState: { type: ["string", "null"] }
        buyerState: { type: ["string", "null"] }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxableAmount: { type: number }
        cgst: { type: number }
        sgst: { type: number }
        igst: { type: number }
        taxRate: { type: number }
        total: { type: number }
        amountPaid: { type: number }
        notes: { type: ["string", "null"] }
        terms: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    InvoiceItem:
      type: object
      required: [id, description, quantity, unitPrice, totalPrice]
      properties:
        id: { type: integer }
        description: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        totalPrice: { type: number }
    InvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/Invoice"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/InvoiceItem" }
            payments:
              type: array
              items: { $ref: "#/components/schemas/Payment" }
            client:
              type: ["object", "null"]
              additionalProperties: true
    InvoiceInput:
      type: object
      properties:
        clientId: { type: ["integer", "null"] }
        salesOrderId: { type: ["integer", "null"] }
        dueDate: { type: ["string", "null"] }
        taxRate: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        terms: { type: ["string", "null"] }
        items:
          type: array
          items:
            type: object
            required: [description, quantity, unitPrice]
            properties:
              description: { type: string }
              quantity: { type: integer }
              unitPrice: { type: number }
    Payment:
      type: object
      required: [id, invoiceId, amount, method, paidAt, createdAt]
      properties:
        id: { type: integer }
        invoiceId: { type: integer }
        amount: { type: number }
        method: { type: string }
        reference: { type: ["string", "null"] }
        paidAt: { type: string }
        notes: { type: ["string", "null"] }
        recordedByName: { type: ["string", "null"] }
        createdAt: { type: string }
...
    Integration:
      type: object
      required: [id, provider, enabled, config, createdAt]
      properties:
        id: { type: integer }
        provider: { type: string }
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: { type: string }
        lastSyncedAt: { type: ["string", "null"] }
        lastSyncStatus: { type: ["string", "null"] }
        lastSyncMessage: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    IntegrationInput:
      type: object
      required: [enabled, config]
      properties:
        enabled: { type: boolean }
        config:
          type: object
          additionalProperties: { type: string }
...