# lib/api-zod/src/generated/types/invoiceDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant SaaS quotation & CRM platform for MSMEs
 * OpenAPI spec version: 2.0.0
 */
import type { Invoice } from './invoice';
import type { InvoiceDetailClient } from './invoiceDetailClient';
import type { InvoiceItem } from './invoiceItem';
import type { Payment } from './payment';

export type InvoiceDetail = Invoice & {
  items?: InvoiceItem[];
  payments?: Payment[];
  /** @nullable */
  client?: InvoiceDetailClient;
};